+++ stegraca-valise/src/components/index.js
@@ -6,5 +6,6 @@ import Details from './details';
 import Footer from './footer';
 import Bio from './biography';
 import Panel from './panel';
+import Linkblock from './linkblock';
 
-export { Header, Valise, Projects, Details, NavBar, Footer, Bio, Panel};
+export { Header, Valise, Projects, Details, NavBar, Footer, Bio, Panel, Linkblock};
