+++ stegraca-valise/src/components/linkblock/index.js
@@ -1,0 +1,1 @@
+export { default } from '../linkblock/Linkblock';
