+++ stegraca-valise/src/components/linkblock/linkblock.css
@@ -1,0 +1,74 @@
+
+.linkblock {
+    grid-area: linkblock;
+    display: grid;
+    grid-template-columns: 1fr 1fr 1fr;
+    grid-template-rows: auto;
+    grid-template-areas: 
+        "about tech portfolio";
+    width: 100%;
+    justify-items: center;
+    position:relative;
+    transition: height 3s;
+}
+
+.tech-icons {
+    width:auto;
+    display: flex;
+    font-size: 4.5cqw;
+    color: white;
+    padding: 0.5rem;
+}
+
+
+
+.tech-icons:hover {
+    color: yellow;
+    stroke: 1px yellowgreen;
+    border-radius: 1rem;
+    transform: scale(1.1);
+    -webkit-transition-duration: 0.4s;
+    -moz-transition-duration: 0.4s;
+    -o-transition-duration: 0.4s;
+    transition-duration: 0.4s;
+}
+
+.tech-used {
+    color:yellow;
+    transform: scale(1.1);
+    -webkit-transition-duration: 0.4s;
+    -moz-transition-duration: 0.4s;
+    -o-transition-duration: 0.4s;
+    transition-duration: 0.4s;
+}
+
+.tech-block {
+    grid-area: tech;
+    display: flex;
+    justify-content: center;
+    width: 100%;
+}
+
+
+.link-text {
+    container-type: inline-size;
+    display: inline-flex;
+    color: white;
+    padding-left: 10px;
+    padding-right: 10px;
+    justify-content:space-between;
+    width:75%;
+}
+
+.page-portfolio {
+    grid-area: portfolio;
+}
+
+.page-about {
+    grid-area: about;
+}
+
+
+.back {
+    grid-area: back;
+}
